Add script to take sceenshots of an ebook.
Add script that will use an android phone to take screenshots of the
ebbok, get the photo to save it to a local disk, and move to the next
page to repeat.

The loop willl continue until the user cancels it, so there is risk of
filling the disk with a lot of duplicates images if we let this run too
far.